+++ Source/CesiumRuntime/Private/CesiumEllipsoidFunctions.cpp
@@ -1,4 +1,4 @@
-// Copyright 2020-2024 CesiumGS, Inc. and Contributors
+// Copyright 2020-2026 CesiumGS, Inc. and Contributors
 
 #include "CesiumEllipsoidFunctions.h"
 #include "VecMath.h"